refactor: extract PDF region block classification
